requirements.txt
torch>=2.0
torchvision>=0.15
PyYAML>=6.0
Pillow>=9.0
numpy>=1.23
tqdm>=4.64
pytest>=7.0
